refactor: Move helpers in database.py to db_utils.py
